Cache the `ref` tag for performance
The `ref` tag is used to link specific skills etc to the relevant
rulebook page. This commit introduces a cache for the `ref` tag to
improve performance.

On a simple 2-fighter list page with no skills, this removes 1 query
because the figher type ref is cached.

On a much more complex list, around 7 duplicate queries are removed
and around 100ms of query time.